fixed migrations

+++ database/migrations/2018_11_11_211017_create_photos_table.php
@@ -13,7 +13,7 @@ class CreatePhotosTable extends Migration {
 	public function up() {
 		Schema::create('photos', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('kitchen_id')->unsigned();
+			$table->bigInteger('kitchen_id')->unsigned();
 			$table->string('file');
 			$table->timestamps();
