+++ database/migrations/2018_11_20_230939_create_products_table.php
@@ -13,7 +13,7 @@ class CreateProductsTable extends Migration {
 	public function up() {
 		Schema::create('products', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('application_id')->unsigned();
+			$table->bigInteger('application_id')->unsigned();
 			$table->string('name');
 			$table->decimal('price');
 			$table->string('category');
