+++ database/migrations/2018_11_24_031941_create_electric_devices_table.php
@@ -13,7 +13,7 @@ class CreateElectricDevicesTable extends Migration {
 	public function up() {
 		Schema::create('electric_devices', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('application_id')->unsigned();
+			$table->bigInteger('application_id')->unsigned();
 			$table->string('name');
 			$table->integer('watts');
 			$table->timestamps();
