+++ database/migrations/2018_12_02_012142_create_invoice_items_table.php
@@ -13,8 +13,8 @@ class CreateInvoiceItemsTable extends Migration {
 	public function up() {
 		Schema::create('invoice_items', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('invoice_id')->unsigned();
-			$table->integer('service_id')->unsigned()->nullable();
+			$table->bigInteger('invoice_id')->unsigned();
+			$table->bigInteger('service_id')->unsigned()->nullable();
 			$table->integer('quantity')->unsigned();
 			$table->integer('tax')->default(0)->unsigned();
 			$table->decimal('unit_price', 10, 2)->unsigned();
